requirements.txt
absl-py==0.9.0
alabaster==0.7.10
aletheia-relunet-unwrapper==1.1.2
anaconda-client==1.6.14
anaconda-navigator==1.9.2
anaconda-project==0.8.2
appdirs==1.4.4
asn1crypto==0.24.0
astor==0.7.1
astroid==1.6.3
astropy==3.0.2
astunparse==1.6.3
atomicwrites==1.4.0
attrs==18.1.0
Babel==2.5.3
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.3
blis==0.4.1
bokeh==0.12.16
boto==2.48.0
boto3==1.13.2
botocore==1.16.2
Bottleneck==1.2.1
Box2D==2.3.10
cachetools==4.1.0
captum==0.2.0
catalogue==1.0.0
certifi==2020.4.5.1
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==1.6.0
clyent==1.2.2
colorama==0.3.9
comtypes==1.1.4
conda==4.8.3
conda-build==3.10.5
conda-package-handling==1.7.0
conda-verify==2.0.0
contextlib2==0.5.5
cryptography==2.9.2
csaps==1.0.3
cycler==0.10.0
cymem==2.0.3
Cython==0.29.14
cytoolz==0.9.0.1
dask==0.17.5
dataclasses==0.7
datashape==0.5.4
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-2.3.0/de_core_news_sm-2.3.0.tar.gz
decorator==4.3.0
defusedxml==0.6.0
distlib==0.3.0
distributed==1.21.8
docutils==0.14
EasyProcess==0.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints==0.2.3
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-2.3.1/es_core_news_sm-2.3.1.tar.gz
et-xmlfile==1.0.1
fastai==1.0.61
fastcache==1.0.2
fastprogress==0.2.3
filelock==3.0.4
Flask==1.0.2
Flask-Cors==3.0.4
future==0.18.2
gast==0.3.3
gensim==3.8.3
gevent==1.3.0
glob2==0.6
google-auth==1.16.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.13.2
greenlet==0.4.13
grpcio==1.29.0
gym==0.18.0
h5py==2.10.0
heapdict==1.0.0
html5lib==1.0.1
idna==2.9
imageio==2.3.0
imagesize==1.0.0
importlib-metadata==1.6.1
importlib-resources==1.5.0
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.12.0
Jinja2==2.10
jmespath==0.9.5
joblib @ file:///tmp/build/80754af9/joblib_1594236160679/work
json5==0.9.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.32.1
jupyterlab-launcher==0.10.5
jupyterlab-server==1.0.0
kaggle==1.5.5
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.1.2
keyring @ file:///C:/ci/keyring_1593109797081/work
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
llvmlite==0.23.1
locket==0.2.0
lxml==4.2.1
Markdown==3.2.2
MarkupSafe==1.0
matplotlib==3.3.4
mccabe==0.6.1
menuinst==1.4.14
mistune==0.8.3
mkl-fft==1.0.0
mkl-random==1.0.1
mkl-service==2.3.0
more-itertools==4.1.0
mpmath==1.0.0
msgpack-python==0.5.6
multipledispatch==0.5.0
murmurhash==1.0.2
navigator-updater==0.2.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.3
nose==1.3.7
notebook==5.5.0
numba==0.38.0
numexpr==2.7.1
numpy==1.18.5
numpydoc==0.8.0
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
odo==0.5.1
olefile==0.45.1
openpyxl==2.5.3
opt-einsum==3.2.1
packaging==17.1
pandas==1.1.5
pandocfilters==1.4.2
parso==0.2.0
partd==0.3.8
path.py==11.0.1
pathlib2==2.3.2
patsy==0.5.0
pep8==1.7.1
pickleshare==0.7.4
Pillow==7.2.0
pkginfo==1.4.2
plac==1.1.3
pluggy==0.6.0
ply==3.11
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==1.0.15
protobuf==3.12.2
psutil==5.4.5
py==1.5.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.5.0
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.5
pydot==1.4.1
pyflakes==1.6.0
pyglet==1.5.0
Pygments==2.2.0
pylint==1.8.4
pyodbc==4.0.23
pyOpenSSL==18.0.0
pyparsing==2.2.0
pyreadline==2.1
pyrsistent==0.16.0
PySocks==1.6.8
pytest==3.5.1
pytest-arraydiff==0.2
pytest-astropy==0.3.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.2.1
python-dateutil==2.7.3
python-slugify==3.0.3
pytz==2018.4
PyVirtualDisplay==2.0
PyWavelets==0.5.2
pywin32==223
pywin32-ctypes==0.2.0
pywinpty==0.5.1
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.1
requests==2.23.0
requests-oauthlib==1.3.0
rope==0.10.7
rsa==4.0
ruamel-yaml==0.15.35
s3transfer==0.3.3
scikit-image==0.13.1
scikit-learn @ file:///C:/ci/scikit-learn_1592853528256/work
scipy==1.4.1
seaborn==0.11.1
selenium==3.141.0
Send2Trash==1.5.0
sentencepiece==0.1.94
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
sklearn==0.0
smart-open==2.0.0
snowballstemmer==1.2.1
sortedcollections==0.6.1
sortedcontainers==1.5.10
soupsieve==2.0.1
spacy==2.3.1
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
spyder==3.2.8
spyder-kernels==1.4.0
SQLAlchemy==1.2.7
srsly==1.0.2
statsmodels==0.12.1
superfast @ file:///E:/visual_studio/CppPythonExtensionExample_Python/superfastcode
superfastcode2 @ file:///E:/visual_studio/CppPythonExtensionExample_Python/superfastcode2
sympy==1.1.1
tables==3.4.3
TBB==0.1
tblib==1.3.2
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
tensorflow-gpu-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
text-unidecode==1.2
thinc==7.4.1
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
toolz==0.9.0
torch==1.7.1
torchtext==0.6.0
torchvision==0.6.0+cu101
tornado==5.0.2
tqdm @ file:///tmp/build/80754af9/tqdm_1593446365756/work
traitlets==4.3.2
typed-ast==1.4.1
typing==3.6.4
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3==1.25.9
virtualenv==20.0.21
wasabi==0.7.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.2.1
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.12.1
xgboost==1.2.0
xlrd==1.1.0
XlsxWriter==1.0.4
xlwings==0.11.8
xlwt==1.3.0
zict==0.1.3
zipp==3.1.0